+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Fixed
 
+* Bug in [HexaPDF::Font::TrueType::Subsetter#use_glyph] which lead to corrupt
+  text output
 * [HexaPDF::Type::Resources] validation to handle an invalid `/ProcSet` entry
   containing a single value instead of an array
 * Processing of invalid PDF files missing a required value in appearance streams